+++ src/Filtering/Filter.ts
@@ -24,7 +24,6 @@ import Recursive from "./Recursive";
  */
 
 interface FilterObj {
-  isstring: boolean;
   regexp: string | RegExp;
   boards: any;
   excludes: any;
@@ -32,8 +31,8 @@ interface FilterObj {
   hide: boolean;
   stub: any;
   hl: string;
-  top: boolean;
-  noti: boolean;
+  top?: boolean;
+  noti?: boolean;
   poster?: boolean;
   replies?: boolean;
   reason?: string;
@@ -47,7 +46,7 @@ export interface FilterResults {
   noti?: boolean;
   poster?: boolean;
   replies?: boolean;
-  reasons?: string;
+  reasons?: string[];
 };
 
 type FilterType = "postID" | "name" | "uniqueID" | "tripcode" | "capcode" | "pass" | "email" | "subject" | "comment"
@@ -71,34 +70,32 @@ var Filter = {
     for (var key in Config.filter) {
       for (var line of (Conf[key] as string).split('\n')) {
         let hl:       string;
-        let isstring: boolean;
         let regexp:   RegExp | string;
         let top:      boolean;
-        let types:    string[];
-        let hide:     boolean;
+        let hide =    true;
+        let mask =    0;
+        let boards:   any = false;
+        let excludes: any = false;
+        let reason:   string | undefined;
+        let poster =  false;
+        let replies = false;
+        let noti =    false;
+        let stub =    Conf.Stubs;
 
         if (line[0] === '#') continue;
 
-        if (!(regexp = line.match(/\/(.*)\/(\w*)/))) {
+        const regexpMatch = line.match(/\/(.*)\/(\w*)/);
+        if (!regexpMatch) {
           continue;
         }
 
-        // Don't mix up filter flags with the regular expression.
-        var filter = line.replace(regexp[0], '');
-
-        // List of the boards this filter applies to.
-        var boards = this.parseBoards(filter.match(/(?:^|;)\s*boards:([^;]+)/)?.[1]);
-
-        // Boards to exclude from an otherwise global rule.
-        var excludes = this.parseBoards(filter.match(/(?:^|;)\s*exclude:([^;]+)/)?.[1]);
-
-        if (isstring = (['uniqueID', 'MD5'].includes(key))) {
+        if (key === 'uniqueID' || key === 'MD5') {
           // MD5 filter will use strings instead of regular expressions.
-          regexp = regexp[1];
+          regexp = regexpMatch[1];
         } else {
           try {
             // Please, don't write silly regular expressions.
-            regexp = RegExp(regexp[1], regexp[2]);
+            regexp = RegExp(regexpMatch[1], regexpMatch[2]);
           } catch (err) {
             // I warned you, bro.
             new Notice('warning', [
@@ -112,60 +109,67 @@ var Filter = {
           }
         }
 
-        // Filter OPs along with their threads or replies only.
-        var op = filter.match(/(?:^|;)\s*op:(no|only)/)?.[1] || '';
-        var mask = $.getOwn({'no': 1, 'only': 2}, op) || 0;
+        // Don't mix up filter flags with the regular expression.
+        const options = line.slice(regexpMatch[0].length);
 
-        // Filter only posts with/without files.
-        var file = filter.match(/(?:^|;)\s*file:(no|only)/)?.[1] || '';
-        mask = mask | ($.getOwn({'no': 4, 'only': 8}, file) || 0);
+        if (options.length) {
 
-        // Overrule the `Show Stubs` setting.
-        // Defaults to stub showing.
-        var stub = (() => { switch (filter.match(/(?:^|;)\s*stub:(yes|no)/)?.[1]) {
-          case 'yes':
-            return true;
-          case 'no':
-            return false;
-          default:
-            return Conf['Stubs'];
-        } })();
+          // List of the boards this filter applies to.
+          boards = this.parseBoards(options.match(/(?:^|;)\s*boards:([^;]+)/)?.[1]);
 
-        // Desktop notification
-        var noti = /(?:^|;)\s*notify/.test(filter);
+          // Boards to exclude from an otherwise global rule.
+          excludes = this.parseBoards(options.match(/(?:^|;)\s*exclude:([^;]+)/)?.[1]);
 
-        // Highlight the post.
-        // If not specified, the highlight class will be filter-highlight.
-        const highlightRes = filter.match(/(?:^|;)\s*highlight(?::([\w-]+))?/)
-        if (highlightRes) {
-          hl = highlightRes[1] || 'filter-highlight';
-          // Put highlighted OP's thread on top of the board page or not.
-          // Defaults to on top.
-          top = (filter.match(/(?:^|;)\s*top:(yes|no)/)?.[1] || 'yes') === 'yes';
-          hide = /(?:^|;)\s*hide(?:[;:]|$)/.test(filter);
+          // Filter OPs along with their threads or replies only.
+          const op = options.match(/(?:^|;)\s*op:(no|only)/)?.[1] || '';
+          mask = $.getOwn({'no': 1, 'only': 2}, op) || 0;
+
+          // Filter only posts with/without files.
+          const file = options.match(/(?:^|;)\s*file:(no|only)/)?.[1] || '';
+          mask = mask | ($.getOwn({'no': 4, 'only': 8}, file) || 0);
+
+          // Overrule the `Show Stubs` setting.
+          // Defaults to stub showing.
+          stub = (() => { switch (options.match(/(?:^|;)\s*stub:(yes|no)/)?.[1]) {
+            case 'yes':
+              return true;
+            case 'no':
+              return false;
+            default:
+              return Conf['Stubs'];
+          } })();
+
+          // Desktop notification
+          noti = /(?:^|;)\s*notify/.test(options);
+
+          // Highlight the post.
+          // If not specified, the highlight class will be filter-highlight.
+          const highlightRes = options.match(/(?:^|;)\s*highlight(?::([\w-]+))?/)
+          if (highlightRes) {
+            hl = highlightRes[1] || 'filter-highlight';
+            // Put highlighted OP's thread on top of the board page or not.
+            // Defaults to on top.
+            top = (options.match(/(?:^|;)\s*top:(yes|no)/)?.[1] || 'yes') === 'yes';
+            hide = /(?:^|;)\s*hide(?:[;:]|$)/.test(options);
+          }
+
+          // Hide the post (default case).
+          hide = hide || !(hl || noti);
+
+          reason = options.match(/(?:^|;)\s*reason:([^;$]+)/)?.[1];
+
+          poster = /(?:^|;)\s*poster(?:[;:]|$)/.test(options);
+
+          replies = /(?:^|;)\s*replies(?:[;:]|$)/.test(options);
         }
+
+        const filterObj: FilterObj
+          = { regexp, boards, excludes, mask, hide, stub, hl, top, noti, reason, poster, replies };
 
         // Fields that this filter applies to (for 'general' filters)
         if (key === 'general') {
-          if (types = filter.match(/(?:^|;)\s*type:([^;]*)/)) {
-            types = types[1].split(',');
-          } else {
-            types = ['subject', 'name', 'filename', 'comment'];
-          }
-        }
-
-        // Hide the post (default case).
-        hide = hide || !(hl || noti);
-
-        const reason = filter.match(/(?:^|;)\s*reason:([^;$]+)/)?.[1];
-
-        const poster = /(?:^|;)\s*poster(?:[;:]|$)/.test(filter);
-
-        const replies = /(?:^|;)\s*replies(?:[;:]|$)/.test(filter);
-
-        const filterObj: FilterObj
-          = { isstring, regexp, boards, excludes, mask, hide, stub, hl, top, noti, reason, poster, replies };
-        if (key === 'general') {
+          const types = options.match(/(?:^|;)\s*type:([^;]*)/)?.[1].split(',')
+            || ['subject', 'name', 'filename', 'comment'];
           for (var type of types) {
             this.filters.get(type)?.push(filterObj) ?? this.filters.set(type, [filterObj]);
           }
@@ -202,7 +206,7 @@ var Filter = {
 
   // Parse comma-separated list of boards.
   // Sites can be specified by a beginning part of the site domain followed by a colon.
-  parseBoards(boardsRaw) {
+  parseBoards(boardsRaw: string) {
     let boards;
     if (!boardsRaw) { return false; }
     if (boards = Filter.parseBoardsMemo[boardsRaw]) { return boards; }
@@ -255,12 +259,14 @@ var Filter = {
         const filtersForType: FilterObj[] = Array.isArray(filtersOrMap) ? filtersOrMap : filtersOrMap.get(value);
         if (!filtersForType) continue;
 
+        const isString = type === 'uniqueID' || type === 'MD5';
+
         for (const filter of filtersForType) {
           if (
             (filter.boards   && !(filter.boards[board]   || filter.boards[site]  )) ||
             (filter.excludes &&  (filter.excludes[board] || filter.excludes[site])) ||
             (filter.mask & mask) ||
-            (filter.isstring ? (filter.regexp !== value) : !filter.regexp.test(value))
+            (isString ? (filter.regexp !== value) : !(filter.regexp as RegExp).test(value))
           ) continue;
           if (filter.hide) {
             if (hideable) {
@@ -271,7 +277,7 @@ var Filter = {
               }
             }
           }
-          if (!hl?.includes(filter.hl)) {
+          if (filter.hl && !hl?.includes(filter.hl)) {
             (hl || (hl = [])).push(filter.hl);
           }
           if (!top) { ({ top } = filter); }
@@ -281,7 +287,7 @@ var Filter = {
         }
       }
     }
-    post.filterResults = {hide, stub, hl, top, noti, poster, replies, reasons: reasons?.join(' & ')};
+    post.filterResults = {hide, stub, hl, top, noti, poster, replies, reasons};
     return post.filterResults;
   },
 
@@ -296,6 +302,32 @@ var Filter = {
       this,
       (!this.isFetchedQuote && (this.isReply || (g.VIEW === 'index')))
     );
+
+    // Add temporary filter for the poster ID for future posts.
+    let reason: string;
+    if (poster && this.info.uniqueID) {
+      reason = `Hidden because it's the same poster as ${this.ID} (${this.filterResults.reasons})`;
+      const { uniqueID } = this.info;
+      const newFilter: FilterObj = {
+        regexp: uniqueID,
+        boards: false,
+        excludes: false,
+        mask: 0,
+        hide,
+        stub,
+        replies,
+        // A filter can only have one hl class.
+        hl: hl?.[0],
+        reason,
+      }
+      const map: Map<string, FilterObj[]> = Filter.filters.get('uniqueID');
+      if (map) {
+        map.get(uniqueID)?.push(newFilter) ?? map.set(uniqueID, [newFilter]);
+      } else {
+        Filter.filters.set('uniqueID', (new Map()).set(uniqueID, [newFilter]))
+      }
+    }
+
     if (hide) {
       if (this.isReply) {
         PostHiding.hide(this, stub);
@@ -303,7 +335,6 @@ var Filter = {
           Recursive.applyAndAdd(PostHiding.hide, this, stub, undefined, `Hidden recursively from ${this.ID}`);
         }
         if (poster && this.info.uniqueID) {
-          const reason = `Hidden because it's the same poster as ${this.ID} (${this.filterResults.reasons})`;
           g.posts.forEach((p) => {
             if (p.info.uniqueID === this.info.uniqueID && p !== this) {
               PostHiding.hide(p, stub, replies, reason);